+++ src/cpp/core/include/core/system/Xdg.hpp
@@ -84,8 +84,9 @@ FilePath oldUserCacheDir(
 
 #endif
 
-// This function verifies that the userConfigDir() and userDataDir() exist and are owned by the running user.
-// 
+// This function verifies that the userConfigDir(), userDataDir(), and the user log directory
+// exist and are writable by the running user.
+//
 // It should be invoked once. Any issues with these directories will be emitted to the session log.
 void verifyUserDirs(const boost::optional<std::string>& user = boost::none,
                     const boost::optional<FilePath>& homeDir = boost::none);